Redesign Invoice Coverletter to fit windowed envelopes
In order to provide the desired functionality we switched our underlying
PDF-rendering engine from `wkhtmltopdf` to `weasyprint`. For several
years now, PRs that would have allowed us to change individual page
orientation have been stale upstream. As this is a required
funtionality, we needed to switch to an engine that provides it.
As a side effect of this change, this commit provides a minor
refactoring of the underlying template which no longer extends the main
(html) 'base-template'.

Ref.: ABSYS-20