Refactor notification formset handling
Previously, the final form was the result of a mix of responsibilities
between django-extra model formsets, custom html and crispy-form classes
for the 'per form' rendering.
This has been streamlined (initially motivated by the fact that the old
per form processing with cripy form broke saving updated model
instances) so that django-extra forms is solely in charge of generating
the required `formset` while `crisply-forms` is handling all form
rendering. The result is that we no longer rely on custom html logic
formset rendering at all. As the entire formset is not under
`crispy-form`'s control, saving updated instances is no longer an issue.

Ref.: ABSYS-57