Remove fixed default for SchuelerInEinrichtung.fehltage_erlaubt
Instead of the fixed default value, we provide access to the computed
value that should probably go here.
However, because at instance creation time (within the form) the
required values are not known, we can not actualle use this as a default
at runtime. This would probably require some rather sofisticated JS
instead.
Maybe we can come back to this and see if we find a way to provide
additional convenience.